+++ src/components/BulkEditList/BulkEditListResult/BulkEditInApp/BulkEditInApp.js
@@ -9,6 +9,7 @@ import {
 import { BulkEditInAppTitle } from './BulkEditInAppTitle/BulkEditInAppTitle';
 import { ItemForm } from './forms/ItemForm/ItemForm';
 import { UserForm } from './forms/UserForm/UserForm';
+import { HoldingsForm } from './forms/HoldingsForm/HoldingsForm';
 import { CAPABILITIES } from '../../../../constants';
 
 export const BulkEditInApp = (
@@ -58,7 +59,7 @@ export const BulkEditInApp = (
         )}
 
         {typeOfBulk === CAPABILITIES.HOLDINGS && (
-          <UserForm
+          <HoldingsForm
             onContentUpdatesChanged={onContentUpdatesChanged}
             typeOfBulk={typeOfBulk}
             getFilteredFields={getFilteredFields}