+++ src/components/BulkEditList/BulkEditListResult/BulkEditInApp/forms/HoldingsForm/HoldingsForm.js
@@ -11,19 +11,15 @@ import {
   RepeatableField,
 } from '@folio/stripes/components';
 
+import { LocationLookup, LocationSelection } from '@folio/stripes/smart-components';
 import {
+  ACTIONS,
   BASE_ACTIONS,
-  USER_OPTIONS,
+  HOLDINGS_OPTIONS, OPTIONS,
 } from '../../../../../../constants';
 import css from '../../BulkEditInApp.css';
-import { ActionsRow } from './ActionsRow';
 
-const fieldsTypes = {
-  ACTION: 'actions',
-  OPTION: 'option',
-};
-
-export const UserForm = (
+export const HoldingsForm = (
   {
     onContentUpdatesChanged,
     getFilteredFields,
@@ -31,50 +27,63 @@ export const UserForm = (
 ) => {
   const intl = useIntl();
 
+  const fieldsTypes = {
+    ACTION: 'action',
+    OPTION: 'option',
+  };
+
   const actions = BASE_ACTIONS(intl.formatMessage);
-  const optionsUser = USER_OPTIONS(intl.formatMessage);
+  const optionsHoldings = HOLDINGS_OPTIONS(intl.formatMessage);
 
   const fieldTemplate = {
-    options: optionsUser,
-    actionsList: actions,
-    option: optionsUser[0].value,
-    actions: getDefaultActions(optionsUser[0].value, intl.formatMessage),
+    options: optionsHoldings,
+    actions,
+    option: optionsHoldings[0].value,
+    action: actions[0].value,
+    value: '',
+    locationId: '',
   };
 
   const [fields, setFields] = useState([fieldTemplate]);
 
+  const isLocation = (index) => fields[index].action === ACTIONS.REPLACE &&
+      (fields[index].option === OPTIONS.PERMANENT_HOLDINGS_LOCATION ||
+          fields[index].option === OPTIONS.TEMPORARY_HOLDINGS_LOCATION);
 
-  const handleOptionChange = (e, index) => {
+  const handleSelectChange = (e, index, type) => {
     const mappedFields = fields.map((field, i) => {
       if (i === index) {
         const value = e.target.value;
 
         return {
           ...field,
-          [fieldsTypes.OPTION]: value,
-          actions: getDefaultActions(value, intl.formatMessage),
+          [type]: value,
         };
       }
 
       return field;
     });
 
-    const recoveredFields = mappedFields.map(f => ({ ...f, options: optionsUser }));
-    const finalizedFields = getFilteredFields(recoveredFields);
+    if (type === fieldsTypes.OPTION) {
+      const recoveredFields = mappedFields.map(f => ({ ...f, optionsHoldings }));
+      const finalizedFields = getFilteredFields(recoveredFields);
 
-    setFields(finalizedFields);
+      setFields(finalizedFields);
+    } else {
+      setFields(mappedFields);
+    }
   };
 
-  const handleActionsChange = ({ rowIndex, actionIndex, actionValue, actionFieldName }) => {
+  const handleLocationChange = (location, index) => {
     setFields(prevFields => prevFields.map((field, i) => {
-      if (i === rowIndex) {
-        return {
-          ...field,
-          actions: field.actions.map((action, j) => (j === actionIndex
-            ? ({ ...action, [actionFieldName]: actionValue })
-            : action
-          )),
-        };
+      if (i === index) {
+        if (location) {
+          return {
+            ...field,
+            locationId: location.id,
+            value: location.name,
+          };
+        }
       }
 
       return field;
@@ -83,23 +92,19 @@ export const UserForm = (
 
   const handleRemove = (index) => {
     const filteredFields = fields.filter((_, i) => i !== index);
-    const recoveredFields = filteredFields.map(f => ({ ...f, options: optionsUser }));
+    const recoveredFields = filteredFields.map(f => ({ ...f, optionsHoldings }));
     const finalizedFields = getFilteredFields(recoveredFields);
 
     setFields(finalizedFields);
   };
 
   const handleAdd = () => {
-    const filteredFields = getFilteredFields([...fields, { ...fieldTemplate, actions: [], option: '' }]);
+    const filteredFields = getFilteredFields([...fields, { ...fieldTemplate, action: '', option: '' }]);
     const initializedFields = filteredFields.map((f, i) => {
       const value = f.options[0].value;
 
       return i === filteredFields.length - 1
-        ? ({
-          ...f,
-          option: value,
-          actions: getDefaultActions(value, intl.formatMessage),
-        })
+        ? ({ ...f, option: value })
         : f;
     });
 
@@ -109,13 +114,7 @@ export const UserForm = (
   };
 
   useEffect(() => {
-    const mappedContentUpdates = fields.map(
-      // eslint-disable-next-line no-shadow
-      ({ option, actions }) => ({
-        option,
-        actions: actions.map(({ name, value }) => ({ name, value })),
-      }),
-    );
+    const mappedContentUpdates = fields.map(({ option, action, value }) => ({ option, action, value }));
 
     onContentUpdatesChanged(mappedContentUpdates);
   }, [fields]);
@@ -131,16 +130,35 @@ export const UserForm = (
             <Select
               dataOptions={field.options}
               value={field.option}
-              onChange={(e) => handleOptionChange(e, index)}
+              onChange={(e) => handleSelectChange(e, index, fieldsTypes.OPTION)}
               data-testid={`select-option-${index}`}
             />
           </Col>
-          <ActionsRow
-            actions={field.actions}
-            onChange={(values) => handleActionsChange({ ...values, rowIndex: index })}
-          />
+          <Col xs={2} sm={2}>
+            <Select
+              dataOptions={field.actions}
+              value={field.action}
+              onChange={(e) => handleSelectChange(e, index, fieldsTypes.ACTION)}
+              data-testid={`select-actions-${index}`}
+            />
+          </Col>
+          {isLocation(index) &&
+          <Col xs={6} sm={3}>
+            <LocationSelection
+              value={field.locationId}
+              onSelect={(location) => handleLocationChange(location, index)}
+              placeholder={intl.formatMessage({ id: 'ui-bulk-edit.layer.selectLocation' })}
+              data-test-id={`textField-${index}`}
+            />
+            <LocationLookup
+              marginBottom0
+              onLocationSelected={(location) => handleLocationChange(location, index)}
+              data-testid={`locationLookup-${index}`}
+            />
+          </Col>
+          }
           <div className={css.iconButtonWrapper}>
-            {(index === fields.length - 1 && fields.length !== optionsUser.length) && (
+            {(index === fields.length - 1 && fields.length !== optionsHoldings.length) && (
             <IconButton
               icon="plus-sign"
               size="large"
@@ -161,7 +179,7 @@ export const UserForm = (
   );
 };
 
-UserForm.propTypes = {
+HoldingsForm.propTypes = {
   onContentUpdatesChanged: PropTypes.func,
   getFilteredFields: PropTypes.func,
 };